requirements.txt
flask
psycopg2-binary
werkzeug
gunicorn
flask-cors
